created steroid class

+++ src/game/Purchasable.java
@@ -4,20 +4,20 @@ package game;
  * abstract class for everything sold at markets in-game
  * 
  * @author Findlay Royds
- * @version 1.2, May 2023.
+ * @version 1.3, May 2023.
  */
 public abstract class Purchasable {
 	/**
 	 * The amount of money the player is charged when purchasing from a market
 	 */
-	protected float price;
+	protected int price;
 	
 	/**
 	 * A constructor for for Purchasable
 	 * 
 	 * @param price					The price of the purchasable
 	 */
-	public Purchasable(float price) {
+	public Purchasable(int price) {
 		this.price = price;
 	}
 	
@@ -32,7 +32,7 @@ public abstract class Purchasable {
 	 * 
 	 * @return 						The price of the purchasable
 	 */
-	public float getPrice() {
+	public int getPrice() {
 		return price;
 	}
 }
